+++ server/src/migrations/006_badges.sql
@@ -1,1 +1,2 @@
-UPDATE badges SET name = 'Compte Certifié' WHERE id = 'v.i.p';
+DELETE FROM user_badges WHERE badgeId = 'v.i.p';
+DELETE FROM badges WHERE id = 'v.i.p';